feat: adding gigahorse client (#17)
## Description
This PR adds a new client based upon gigahorse so that users can have a
sonatype client which supports Java 8. Testing for Java 8 will require a
separate testing module specific to gigahorse, since sttp is failing
even to compile on Java 8 with Scala 3. This will need to be introduced
in a separate PR.

/closes #16